+++ server/src/main/java/net/feedbacky/app/data/board/dto/FetchBoardDto.java
@@ -28,6 +28,8 @@ public class FetchBoardDto implements FetchResponseDto<FetchBoardDto, Board> {
   private String fullDescription;
   private long creatorId;
   private boolean anonymousAllowed;
+  private boolean roadmapEnabled;
+  private boolean changelogEnabled;
   private Date creationDate;
 
   private List<FetchSocialLinkDto> socialLinks;
@@ -60,6 +62,8 @@ public class FetchBoardDto implements FetchResponseDto<FetchBoardDto, Board> {
     this.fullDescription = entity.getFullDescription();
     this.creatorId = entity.getCreator().getId();
     this.anonymousAllowed = entity.isAnonymousAllowed();
+    this.roadmapEnabled = entity.isRoadmapEnabled();
+    this.changelogEnabled = entity.isChangelogEnabled();
     this.creationDate = entity.getCreationDate();
     this.socialLinks = entity.getSocialLinks().stream().map(link -> new FetchSocialLinkDto().from(link)).collect(Collectors.toList());
     this.tags = entity.getTags().stream().map(tag -> new FetchTagDto().from(tag)).collect(Collectors.toList());